--- priv/react_app/src/Screens/Admin/StudentSession/StudentSession.test.js
@@ -1,55 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import { Table } from 'antd';
-
-import StudentSession from './StudentSession';
-import LoadingSpinner from '../../../Components/LoadingSpinner';
-
-describe('companies', () => {
-  let props;
-  beforeEach(() => {
-    props = {
-      fetching: false,
-      getAllCompanies: jest.fn(),
-      deleteCompany: jest.fn()
-    };
-  });
-
-  it('should render without crashing', () => {
-    shallow(<StudentSession {...props} />);
-  });
-
-  it('should render loading', () => {
-    const wrapper = shallow(<StudentSession {...props} fetching />);
-    expect(wrapper.find(LoadingSpinner)).toHaveLength(1);
-  });
-
-  it('calls fetch all companies prop on mount', () => {
-    const func = jest.fn();
-    shallow(<StudentSession {...props} getAllCompanies={func} />);
-
-    expect(func).toHaveBeenCalledTimes(1);
-  });
-
-  it('should render companies', () => {
-    const companies = {
-      '1': {
-        id: 1,
-        name: 'Spotify',
-        description: 'We do music!',
-        website: 'www.spotify.com'
-      },
-      '2': {
-        id: 2,
-        name: 'Google',
-        description: 'We code!',
-        website: 'www.google.com'
-      }
-    };
-    const wrapper = shallow(
-      <StudentSession id="1" {...props} companies={companies} />
-    );
-
-    expect(wrapper.find(Table)).toHaveLength(1);
-  });
-});
