+++ priv/react_app/src/Screens/Admin/StudentSessions/index.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import StudentSessions from './StudentSessions';
+import { Actions } from '../../../Store';
+
+const mapDispatchToProps = {
+  createBulkStudentSessions: Actions.studentSessions.createBulkStudentSessions
+};
+
+const stateful = connect(
+  null,
+  mapDispatchToProps
+);
+
+export default stateful(StudentSessions);
