--- priv/react_app/src/Screens/YourCompany/YourCompanySchedule/YourCompanySchedule.test.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import YourCompanySchedule from './YourCompanySchedule';
-
-it('should render without crashing', () => {
-  shallow(<YourCompanySchedule />);
-});
