--- priv/react_app/src/Screens/YourCompany/YourCompanySchedule/index.js
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux';
-import { Actions, Selectors } from '../../../Store';
-import YourCompanySchedule from './YourCompanySchedule';
-import { State } from '../../../Store/reducers/index';
-
-const mapStateToProps = (state: State) => ({
-  currentCompany: Selectors.companies.getCurrentCompany(state),
-  fetching: state.api.currentCompany.fetching
-});
-
-const mapDispatchToProps = {
-  getCurrentCompany: Actions.companies.getCurrentCompany
-  // destroyCurrentCompany: Actions.companies.destroyCurrentCompany
-};
-
-const stateful = connect(
-  mapStateToProps,
-  mapDispatchToProps
-);
-
-export default stateful(YourCompanySchedule);
